blacklists.py: All blacklists are now actually lists
Blacklist class refactored to inherit from and extend the built-in list object.

Define a number of supporting classes to enable this behavior:

WatchMixin provides behavior for watch classes.

PhoneMixin contains beavior specific to phone number lists.

WhoWhatWhenString subclasses str, provides a string with additional attributes for TSVDictParser and subclasses (watchlists with information about who watched something and when).

In Blacklist and  BlacklistParser, rename the remove method to delete, so as not to shadow the list superclass method

The global blacklist variables now contain the actual list objects, not merely a list of strings.

findspam.py: watched_keywords is now a list